start refactoring to do as much text processing as possible at load time
The essential change that this starts to do is make it so that all of the
processFoo functions return a function that takes Path, Substitutions, and
Library (of those, possibly only the substitutions actually needs to be deferred
to here?). And now, as much computation is done _before_ producing that
function. Anything that happens _inside_ that function will happen at
render-time.

This should seem similar to how Compiled Heist works, because indeed, Larceny
is supposed to be a middle-ground between Interpreted Heist and Compiled Heist.