Convert XML to a Larceny type to support namespaces
This will allow us to do a little processing when the template is
loaded, before the template is actually rendered -- I think it's
especially important for parsing the namespaces and attributes.
Instead of that parsing happening over and over again on every
render, it happens once, when the app is started and the templates
loaded.

This also adds support for namespaces. (I tried to make it
two commits but failed.)